requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
anthropic==0.91.0
python-dotenv==1.2.2
PyGithub==2.9.0
python-multipart==0.0.24
